Moved enum to app/enums
