+++ src/app/enums/acl-flags.enum.ts
@@ -1,0 +1,8 @@
+export enum AclFlag {
+  FileInherit = 'FILE_INHERIT',
+  DirectoryInherit = 'DIRECTORY_INHERIT',
+  NoPropagateInherit = 'NO_PROPAGATE_INHERIT',
+  InheritOnly = 'INHERIT_ONLY',
+  Inherited = 'INHERITED',
+  Basic = 'BASIC',
+}
