+++ src/app/interfaces/acl-flags.interface.ts
@@ -1,10 +1,3 @@
-export enum AclFlag {
-  FileInherit = 'FILE_INHERIT',
-  DirectoryInherit = 'DIRECTORY_INHERIT',
-  NoPropagateInherit = 'NO_PROPAGATE_INHERIT',
-  InheritOnly = 'INHERIT_ONLY',
-  Inherited = 'INHERITED',
-  Basic = 'BASIC',
-}
+import { AclFlag } from 'app/enums/acl-flags.enum';
 
 export type AclFlags = { [K in AclFlag]: boolean };
